+++ rest-openehr/src/test/java/org/ehrbase/rest/openehr/OpenehrTemplateControllerTest.java
@@ -23,6 +23,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.spy;
+import static org.mockito.Mockito.when;
 
 import com.nedap.archie.rm.composition.Composition;
 import java.util.List;
@@ -31,7 +32,6 @@ import org.ehrbase.api.service.CompositionService;
 import org.ehrbase.api.service.TemplateService;
 import org.ehrbase.openehr.sdk.response.dto.ehrscape.StructuredString;
 import org.ehrbase.openehr.sdk.response.dto.ehrscape.StructuredStringFormat;
-import org.ehrbase.openehr.sdk.response.dto.ehrscape.TemplateMetaDataDto;
 import org.ehrbase.openehr.sdk.webtemplate.model.WebTemplate;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -70,9 +70,9 @@ class OpenehrTemplateControllerTest {
     }
 
     private OpenehrTemplateController controller() {
-        doReturn(SAMPLE_ID).when(mockTemplateService).create(any());
-        doReturn(SAMPLE_OPT).when(mockTemplateService).findOperationalTemplate(any(), any());
-        doReturn(SAMPLE_WEB_TEMPLATE).when(mockTemplateService).findWebTemplate(SAMPLE_ID);
+        when(mockTemplateService.create(any())).thenReturn(SAMPLE_ID);
+        when(mockTemplateService.findOperationalTemplate(any())).thenReturn(SAMPLE_OPT);
+        when(mockTemplateService.findWebTemplate(SAMPLE_ID)).thenReturn(SAMPLE_WEB_TEMPLATE);
         return spyController;
     }
 
@@ -80,24 +80,25 @@ class OpenehrTemplateControllerTest {
     @CsvSource({"application/json", "application/xml"})
     void getTemplatesADL1_4(String accept) {
 
-        TemplateMetaDataDto metaDataDto = new TemplateMetaDataDto();
-        metaDataDto.setTemplateId(SAMPLE_ID);
+        var templateDetails = new TemplateService.TemplateDetails(null, SAMPLE_ID, null, null, null);
 
-        doReturn(List.of(metaDataDto)).when(mockTemplateService).getAllTemplates();
+        when(mockTemplateService.findAllTemplates()).thenReturn(List.of(templateDetails));
 
-        var response = controller().getTemplatesClassic("1.0.3", null, accept);
+        var response = controller().getTemplatesClassic(accept);
         assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
         assertThat(response.getHeaders()).containsEntry(HttpHeaders.CONTENT_TYPE, List.of(accept));
         assertThat(response.getHeaders())
                 .containsEntry(HttpHeaders.LOCATION, List.of(CONTEXT_PATH + "/definition/template/adl1.4"));
-        assertThat(response.getBody()).isEqualTo(List.of(metaDataDto));
+        assertThat(response.getBody())
+                .hasSize(1)
+                .allMatch(m -> m.getTemplateId().equals(SAMPLE_ID));
     }
 
     @ParameterizedTest
     @ValueSource(strings = {"", "return=minimal", "return=representation"})
     void createTemplateADL1_4(String prefer) {
 
-        var response = controller().createTemplateClassic("1.0.3", null, prefer, SAMPLE_OPT);
+        var response = controller().createTemplateClassic(prefer, SAMPLE_OPT);
         assertThat(response.getStatusCode()).isEqualTo(HttpStatus.CREATED);
         assertThat(response.getHeaders())
                 .containsEntry(HttpHeaders.CONTENT_TYPE, List.of(MediaType.APPLICATION_XML_VALUE));
@@ -116,15 +117,14 @@ class OpenehrTemplateControllerTest {
     void createTemplateADL1_4_OPTInvalidError() {
 
         OpenehrTemplateController controller = controller();
-        assertThatThrownBy(() -> controller.createTemplateClassic("1.0.3", null, null, "not a xml"))
+        assertThatThrownBy(() -> controller.createTemplateClassic(null, "not a xml"))
                 .isInstanceOf(InvalidApiParameterException.class);
     }
 
     @Test
     void getTemplateADL1_4_OPT() {
 
-        ResponseEntity<?> response =
-                controller().getTemplateClassic("1.0.3", null, MediaType.APPLICATION_XML_VALUE, SAMPLE_ID);
+        ResponseEntity<?> response = controller().getTemplateClassic(MediaType.APPLICATION_XML_VALUE, SAMPLE_ID);
         assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
         assertThat(response.getHeaders()).containsEntry(HttpHeaders.CONTENT_TYPE, List.of("application/xml"));
         assertThat(response.getHeaders())
@@ -137,7 +137,7 @@ class OpenehrTemplateControllerTest {
     @CsvSource({"application/json", "application/openehr.wt+json"})
     void getTemplateADL1_4_WebTemplate(String accept) {
 
-        ResponseEntity<?> response = controller().getTemplateClassic("1.0.3", null, accept, SAMPLE_ID);
+        ResponseEntity<?> response = controller().getTemplateClassic(accept, SAMPLE_ID);
         assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
         assertThat(response.getHeaders()).containsEntry(HttpHeaders.CONTENT_TYPE, List.of(accept));
         assertThat(response.getHeaders())
@@ -191,11 +191,9 @@ class OpenehrTemplateControllerTest {
     void templateADL2NotImplemented() {
 
         OpenehrTemplateController controller = controller();
-        assertThat(controller.getTemplatesNew(null, null, null).getStatusCode()).isEqualTo(HttpStatus.NOT_IMPLEMENTED);
-        assertThat(controller
-                        .createTemplateNew(null, null, null, null, null, null, null)
-                        .getStatusCode())
+        assertThat(controller.getTemplatesNew(null).getStatusCode()).isEqualTo(HttpStatus.NOT_IMPLEMENTED);
+        assertThat(controller.createTemplateNew(null, null, null, null, null).getStatusCode())
                 .isEqualTo(HttpStatus.NOT_IMPLEMENTED);
-        assertThat(controller.getTemplatesNew(null, null, null).getStatusCode()).isEqualTo(HttpStatus.NOT_IMPLEMENTED);
+        assertThat(controller.getTemplatesNew(null).getStatusCode()).isEqualTo(HttpStatus.NOT_IMPLEMENTED);
     }
 }