+++ service/src/main/java/org/ehrbase/service/ValidationServiceImp.java
@@ -41,8 +41,10 @@ import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import org.apache.commons.collections4.CollectionUtils;
 import org.ehrbase.api.exception.InternalServerException;
+import org.ehrbase.api.exception.ObjectNotFoundException;
 import org.ehrbase.api.exception.UnprocessableEntityException;
 import org.ehrbase.api.exception.ValidationException;
+import org.ehrbase.api.service.TemplateService;
 import org.ehrbase.api.service.ValidationService;
 import org.ehrbase.openehr.sdk.response.dto.ContributionCreateDto;
 import org.ehrbase.openehr.sdk.terminology.openehr.TerminologyService;
@@ -72,7 +74,7 @@ public class ValidationServiceImp implements ValidationService {
 
     private static final Pattern NAMESPACE_PATTERN = Pattern.compile("[a-zA-Z][a-zA-Z0-9-_:/&+?]*");
 
-    private final KnowledgeCacheServiceImp knowledgeCacheService;
+    private final TemplateService templateService;
 
     private final TerminologyService terminologyService;
     private final boolean folderValidationEnabled;
@@ -82,12 +84,12 @@ public class ValidationServiceImp implements ValidationService {
     private final Map<String, RMPathQuery> rmPathQueryCache = new ConcurrentHashMap<>();
 
     public ValidationServiceImp(
-            KnowledgeCacheServiceImp knowledgeCacheService,
+            TemplateService templateService,
             TerminologyService terminologyService,
             ValidationProperties validationProperties,
             ObjectProvider<ExternalTerminologyValidation> objectProvider,
             @Value("${cache.validation.useSharedRMPathQueryCache:true}") boolean sharedAqlQueryCache) {
-        this.knowledgeCacheService = knowledgeCacheService;
+        this.templateService = templateService;
         this.terminologyService = terminologyService;
         this.folderValidationEnabled = validationProperties.validateFolders();
 
@@ -167,9 +169,9 @@ public class ValidationServiceImp implements ValidationService {
     private void check(String templateID, Composition composition) {
         WebTemplate webTemplate;
         try {
-            webTemplate = knowledgeCacheService.getInternalTemplate(templateID);
-        } catch (IllegalArgumentException e) {
-            throw new UnprocessableEntityException(e.getMessage());
+            webTemplate = templateService.getInternalTemplate(templateID);
+        } catch (ObjectNotFoundException e) {
+            throw new UnprocessableEntityException(e.getMessage(), e);
         }
 
         // Validate the composition based on WebTemplate