+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.31.0]
+ ### Added
+ ### Changed
+- Improved template cache configuration [#1618](https://github.com/ehrbase/ehrbase/pull/1618)
+- Improved performance of template service [#1618](https://github.com/ehrbase/ehrbase/pull/1618)
+ ### Fixed 
+
 ## [2.30.1]
  ### Added
  ### Changed 
@@ -287,3 +294,4 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 [2.29.0]: https://github.com/ehrbase/ehrbase/compare/v2.28.0...v2.29.0
 [2.30.0]: https://github.com/ehrbase/ehrbase/compare/v2.29.0...v2.30.0
 [2.30.1]: https://github.com/ehrbase/ehrbase/compare/v2.30.0...v2.30.1
+[2.31.0]: https://github.com/ehrbase/ehrbase/compare/v2.30.1...v2.31.0